fix layout rework
