+++ .eslintrc.json
@@ -12,7 +12,7 @@
 			"alias": {
 				"map": [
 					["@component", "./src/component"],
-					["@halaman", "./src/halaman"],
+					["@pages", "./src/pages"],
 					["@hooks", "./src/hooks"],
 					["@config", "./src/config"],
 					["@services", "./src/serveces"],
