+++ src/App.js
@@ -7,10 +7,10 @@ import { Switch, Route, Redirect } from 'react-router-dom';
 import Stunting from '@components/Content/Stunting/Stunting';
 import NewsList from '@components/Content/News/NewsList';
 import GiziList from '@components/Content/Gizi/GiziList';
-import NotFound from '@halaman/NotFound/NotFound';
-import About from '@halaman/About/About';
+import NotFound from '@pages/NotFound/NotFound';
+import About from '@pages/About/About';
 import MpasiList from '@components/Content/Mpasi/MpasiList';
-import Home from '@halaman/Home/Home';
+import Home from '@pages/Home/Home';
 import { TemplateDetailNews } from './components/Templates/TemplateDetailNews';
 import { TemplateDetailGizi } from './components/Templates/TemplateDetailGizi';
 import { TemplateDetailMpasi } from './components/Templates/TemplateDetailMpasi';
