--- src/components/Content/Gizi/GiziItem.module.css
@@ -1,9 +1,0 @@
-.container-img {
-	object-fit: cover;
-	min-width: 250px;
-	min-height: 250px;
-	max-width: 250px;
-	max-height: 250px;
-	border-radius: 8px;
-	margin: 0.5rem;
-}
