+++ src/components/Content/News/NewsList.js
@@ -38,24 +38,27 @@ const NewsList = () => {
 					</ErrorModal>
 				</div>
 			)}
-			<div className={classes['wrapper']}>
+			<div className={classes['font-title']}>
 				<h1>Berita Terkini</h1>
 				<hr />
-				{isLoading && (
-					<div className="centered">
-						<LoadingSpinner />
-					</div>
-				)}
-				{articles.map((data) => (
-					<NewsItem
-						key={data.id}
-						id={data.id}
-						title={data.title}
-						description={data.description}
-						url={data.url}
-						urlToImage={data.urlToImage}
-					/>
-				))}
+
+				<div className={classes['wrapper']}>
+					{isLoading && (
+						<div className="centered">
+							<LoadingSpinner />
+						</div>
+					)}
+					{articles.map((data) => (
+						<NewsItem
+							key={data.id}
+							id={data.id}
+							title={data.title}
+							description={data.description}
+							url={data.url}
+							urlToImage={data.urlToImage}
+						/>
+					))}
+				</div>
 			</div>
 		</div>
 	);
