+++ src/components/Content/News/NewsList.module.css
@@ -1,5 +1,12 @@
 .wrapper {
-	display: block;
+	display: flex;
+	flex-wrap: wrap;
+	padding: 2rem 5rem;
+	gap: 3rem;
+	justify-content: space-between;
+}
+
+.font-title {
 	padding: 2rem 5rem;
 }
