+++ src/components/Content/Products/ProductsList.module.css
@@ -1,5 +1,6 @@
 .wrapper {
 	padding: 4rem;
 	display: flex;
-	justify-content: space-around;
+	flex-wrap: wrap;
+	justify-content: center;
 }
