+++ src/components/Header/Header.module.css
@@ -8,6 +8,8 @@
 	padding-top: 2rem;
 	padding-right: 5rem;
 	padding-left: 5rem;
+	position: fixed;
+	z-index: 1000;
 }
 
 .navbar {
@@ -26,7 +28,6 @@
 	font-weight: 700;
 	display: flex;
 }
-
 
 .navbar_list .navbar_item a:hover {
 	text-decoration: none;
