+++ src/components/UI/ErrorModal.js
@@ -7,11 +7,14 @@ const ErrorModal = React.memo((props) => {
 		<>
 			<div className={classes.backdrop} onClick={props.onClose} />
 			<div className={classes['error-modal']}>
-				<h2>An Error Occurred!</h2>
+				<h2>Terjadi Kesalahan!</h2>
 				<p>{props.children}</p>
 				<div className={classes['error-modal__actions']}>
-					<button type="button" onClick={props.onClose}>
-						Okay
+					<button
+						onClick={props.onClose}
+						className="bg-someting-green hover:bg-someting-dark-green text-white font-bold py-2 px-4 rounded my-4"
+					>
+						Tutup
 					</button>
 				</div>
 			</div>
