--- src/halaman/About/About.css
@@ -1,180 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Poppins:wght@200;300;400&display=swap');
-* {
-	margin: 0px;
-	padding: 0px;
-	box-sizing: border-box;
-	font-family: 'Poppins', sans-serif;
-}
-
-.about-section {
-	width: 100%;
-	min-height: 60vh;
-	/*background-color: #ddd;*/
-}
-
-.container-about {
-	width: 80%;
-	display: block;
-	margin: auto;
-	padding-top: 100px;
-}
-
-.content-section {
-	float: left;
-	width: 55%;
-}
-
-.image-section {
-	float: right;
-	width: 40%;
-}
-
-.image-section img {
-	width: 100%;
-	height: auto;
-}
-
-.content-section .title h1 {
-	text-transform: uppercase;
-	font-size: 45px;
-	font-weight: 700;
-	color: #00525f;
-}
-
-.content-section .content h2 {
-	margin-top: 20px;
-	color: #00525f;
-	font-size: 21px;
-}
-
-.content-section .content p {
-	margin-top: 10px;
-	font-family: sans-serif;
-	font-size: 18px;
-	line-height: 1.5;
-}
-
-@media screen and (max-width: 768px) {
-	.container {
-		width: 80%;
-		display: block;
-		margin: auto;
-		padding-top: 50px;
-	}
-
-	.content-section {
-		float: none;
-		width: 100%;
-		display: block;
-		margin: auto;
-	}
-
-	.image-section {
-		float: none;
-		width: 100%;
-	}
-
-	.image-section img {
-		width: 100%;
-		height: auto;
-		display: block;
-		margin: auto;
-	}
-
-	.content-section .title {
-		text-align: center;
-		font-size: 19px;
-	}
-}
-
-.team {
-	width: 100%;
-}
-
-.team .container-team {
-	width: 80%;
-	margin: auto;
-}
-
-.team .container-team .row {
-	width: 60%;
-	margin: auto;
-}
-
-.team .container-team .row .title-team h3 {
-	font-size: 35px;
-	text-transform: uppercase;
-	letter-spacing: 1.5px;
-	color: #00525f;
-	text-align: center;
-	font-weight: 700;
-}
-
-.team .container-team .team-card {
-	width: 100%;
-	display: grid;
-	grid-template-columns: repeat(5, 1fr);
-	grid-gap: 20px;
-	margin-top: 40px;
-	margin-bottom: 50px;
-}
-
-.team .container-team .team-card .card {
-	width: 100%;
-	background-color: #c9d9db;
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	flex-direction: column;
-	padding: 15px;
-	cursor: pointer;
-}
-
-.team .container-team .team-card .card:hover {
-	background-color: #afc1c4;
-	transition: 0.2s;
-}
-
-.team .container-team .team-card .card .avatar img {
-	height: 150px;
-	width: 150px;
-	border-radius: 50%;
-	border: 5px solid #fff;
-}
-
-.team .container-team .team-card .card .content {
-	width: 100%;
-	text-align: center;
-}
-
-.team .container-team .team-card .card .content h4 {
-	font-size: 20px;
-	margin-top: 10px;
-	letter-spacing: 1.5px;
-	color: #75868a;
-}
-
-.team .container-team .team-card .card .content a {
-	font-style: normal;
-	color: #00525f;
-	font-weight: 700;
-	text-decoration: none;
-}
-
-@media screen and (max-width: 1230px) {
-	.team .container-team .team-card {
-		grid-template-columns: repeat(3, 1fr);
-	}
-}
-
-@media screen and (max-width: 700px) {
-	.team .container-team .team-card {
-		grid-template-columns: repeat(2, 1fr);
-	}
-}
-
-@media screen and (max-width: 500px) {
-	.team .container-team .team-card {
-		grid-template-columns: 1fr;
-	}
-}
