+++ src/pages/AllNews.js
